updating to runtime load dependencies and make backward compatile adding monkeypatch for WBEMConnections   - included a pdb.set_trace() to verify operation.  Then removing   - using "global pywbem" to check for package/module vs try:except
